refactor(coap): move COwn into new helpers module
Small enhancements applied due to issues raised from this:

* The comparison between the pool and the generatable count is now
  explicit; the latter is also reused instead of a constant.
* Field access is replaced with the recently introduced `.as_slice()`
  method.
